Add debug message to parser.
