+++ sources/mandatory/parser/parse.c
@@ -6,74 +6,57 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/09 16:23:33 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/09/13 13:38:42 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/10/14 13:55:16 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-static void	remove_redirection(char **tokens)
+static void	debug_node(t_parse_node *node)
 {
-	ft_strarr_cutn(tokens, 2);
-}
+	t_parse	*parse;
 
-static void	parse_redirections(t_parse_list **list, char **tokens)
-{
-	char	*next_token;
-
-	while (tokens != NULL && *tokens != NULL)
+	parse = node->content;
+	ft_printf("\t{ TYPE: %d, ", parse->type);
+	if (parse->type == PT_EXEC)
 	{
-		if (!is_operator(*tokens))
-		{
-			tokens++;
-			continue ;
-		}
-		if (ft_streq(*tokens, PIPE))
-			break ;
-		next_token = *(tokens + 1);
-		if (ft_streq(*tokens, TRUNCATE))
-			add_truncate(list, next_token);
-		if (ft_streq(*tokens, APPEND))
-			add_append(list, next_token);
-		if (ft_streq(*tokens, READ_FILE))
-			add_read_file(list, next_token);
-		if (ft_streq(*tokens, HEREDOC))
-			add_heredoc(list, next_token);
-		remove_redirection(tokens);
+		ft_printf("TOKENS: { ");
+		ft_put_strarr_inline(parse->tokens);
+		ft_printf(" }");
 	}
+	if (parse->type == PT_TRUNCATE
+		|| parse->type == PT_APPEND
+		|| parse->type == PT_READ_FILE)
+		ft_printf("FILE_PATH: %s", parse->file_path);
+	if (parse->type == PT_HEREDOC)
+		ft_printf("DELIMITER: %s", parse->delimiter);
+	ft_printf(" }\n");
 }
 
-static void	parse_exec(t_parse_list **list, char **tokens)
+static void	debug_list(t_parse_node *node)
 {
-	parse_redirections(list, tokens);
-	add_exec(list, tokens);
-}
-
-static void	parse_tokens(t_parse_list **list, char **tokens)
-{
-	int	pipe_count;
-
-	pipe_count = count_pipes(tokens);
-	while (pipe_count > 0)
+	if (!debug())
+		return ;
+	if (node == NULL)
 	{
-		parse_exec(list, tokens);
-		add_pipe(list);
-		tokens = find_next_pipe(tokens) + 1;
-		pipe_count--;
+		ft_debug("parse = NULL");
+		return ;
 	}
-	parse_exec(list, tokens);
+	ft_yellowb("DEBUG:\t");
+	ft_putstr("parse = [\n");
+	while (node != NULL)
+	{
+		debug_node(node);
+		node = node->next;
+	}
+	ft_putstr("]\n");
 }
 
 t_parse_list	*parse(char **tokens)
 {
 	t_parse_list	*list;
-	char			**free_me;
 
-	if (tokens == NULL)
-		return (NULL);
-	list = NULL;
-	free_me = ft_strarr_dup_shallow(tokens);
-	parse_tokens(&list, free_me);
-	free(free_me);
+	list = parse_tokens(tokens);
+	debug_list(list);
 	return (list);
 }
